优化

+++ app/src/main/java/com/work/guaishouxingqiu/aboutball/other/OSSRequestHelp.java
@@ -121,7 +121,7 @@ public class OSSRequestHelp {
      *
      * @param filePath
      */
-    public void uploadingFile(String filePath, OnOSSResultListener listener) {
+    public void uploadingFile(String filePath, OnOSSResultListener listener, BaseActivity activity) {
 
         //super.run();
         if (filePath == null) {
@@ -156,36 +156,40 @@ public class OSSRequestHelp {
         mOssClient.asyncPutObject(request, new OSSCompletedCallback<PutObjectRequest, PutObjectResult>() {
             @Override
             public void onSuccess(PutObjectRequest request, PutObjectResult result) {
-                //拼接阿里路径
-                String ossFilePath = "https://" + OSSRequestHelp.BUCKET.concat(".").concat(OSSRequestHelp.ENDPOINT_BODY).concat("/")
-                        .concat(request.getObjectKey());
-                LogUtils.w("onSuccess--", "上传成功啦！" + ossFilePath + "--" + Thread.currentThread().getName());
-                FileUtils.removeFile(file);
-                mMainHandler.post(() -> {
-                    if (listener != null) {
-                        listener.onSucceed(ossFilePath);
-                    }
-                });
+                if (activity != null && !activity.isFinishing()) {
+                    //拼接阿里路径
+                    String ossFilePath = "https://" + OSSRequestHelp.BUCKET.concat(".").concat(OSSRequestHelp.ENDPOINT_BODY).concat("/")
+                            .concat(request.getObjectKey());
+                    LogUtils.w("onSuccess--", "上传成功啦！" + ossFilePath + "--" + Thread.currentThread().getName());
+                    FileUtils.removeFile(file);
+                    mMainHandler.post(() -> {
+                        if (listener != null) {
+                            listener.onSucceed(ossFilePath);
+                        }
+                    });
+                }
             }
 
             @Override
             public void onFailure(PutObjectRequest request, ClientException clientException, ServiceException serviceException) {
                /* LogUtils.w("onSuccess--", "上传失败啦！" + request.getObjectKey() + "--" + serviceException.getErrorCode() + "--"
                         + serviceException.getStatusCode());*/
-                mMainHandler.post(() -> {
-                    UIUtils.showToast(R.string.uploading_fail_again);
-                    if (listener != null) {
-                        String errCode = "上传失败";
-                        if (serviceException != null) {
-                            errCode = serviceException.getErrorCode();
-                        } else {
-                            if (clientException != null) {
-                                errCode = clientException.getMessage();
+                if (activity != null && !activity.isFinishing()) {
+                    mMainHandler.post(() -> {
+                        UIUtils.showToast(R.string.uploading_fail_again);
+                        if (listener != null) {
+                            String errCode = "上传失败";
+                            if (serviceException != null) {
+                                errCode = serviceException.getErrorCode();
+                            } else {
+                                if (clientException != null) {
+                                    errCode = clientException.getMessage();
+                                }
                             }
+                            listener.onFailure(errCode);
                         }
-                        listener.onFailure(errCode);
-                    }
-                });
+                    });
+                }
 
             }
         });
@@ -208,7 +212,7 @@ public class OSSRequestHelp {
         return files;
     }
 
-    public void uploadingFiles(List<String> filePathData, OnOSSDataResultListener listener) {
+    public void uploadingFiles(List<String> filePathData, OnOSSDataResultListener listener, BaseActivity activity) {
         if (mFileData == null) {
             mFileData = checkFiles(filePathData);
             if (mFileData.size() == 0) {
@@ -219,7 +223,7 @@ public class OSSRequestHelp {
                 listener.onStart();
             }
         }
-        if (mFileData.size() == 0){
+        if (mFileData.size() == 0) {
             UIUtils.showToast("上传文件不能为空！");
             return;
         }
@@ -238,47 +242,51 @@ public class OSSRequestHelp {
         mOssClient.asyncPutObject(request, new OSSCompletedCallback<PutObjectRequest, PutObjectResult>() {
             @Override
             public void onSuccess(PutObjectRequest request, PutObjectResult result) {
-                mMainHandler.post(() -> {
-                    String ossFilePath = "https://" + OSSRequestHelp.BUCKET.concat(".").concat(OSSRequestHelp.ENDPOINT_BODY).concat("/")
-                            .concat(request.getObjectKey());
-                    mResultPathData.add(ossFilePath);
-                    filePathData.remove(0);
-                    mFileData.remove(0);
-                    FileUtils.removeFile(file);
-                    if (filePathData.size() == 0) {
+                if (activity != null && !activity.isFinishing()) {
+                    mMainHandler.post(() -> {
+                        String ossFilePath = "https://" + OSSRequestHelp.BUCKET.concat(".").concat(OSSRequestHelp.ENDPOINT_BODY).concat("/")
+                                .concat(request.getObjectKey());
+                        mResultPathData.add(ossFilePath);
+                        filePathData.remove(0);
+                        mFileData.remove(0);
+                        FileUtils.removeFile(file);
+                        if (filePathData.size() == 0) {
 
-                        if (listener != null) {
-                            listener.onSucceed(mResultPathData);
+                            if (listener != null) {
+                                listener.onSucceed(mResultPathData);
+                            }
+                        } else {
+                            uploadingFiles(filePathData, listener, activity);
                         }
-                    } else {
-                        uploadingFiles(filePathData, listener);
-                    }
-                });
+                    });
+                }
             }
 
             @Override
             public void onFailure(PutObjectRequest request, ClientException clientException, ServiceException serviceException) {
-                mMainHandler.post(() -> {
-                    UIUtils.showToast(R.string.uploading_fail_again);
-                    if (listener != null) {
-                        String errCode = "上传失败";
-                        if (serviceException != null) {
-                            errCode = serviceException.getErrorCode();
-                        } else {
-                            if (clientException != null) {
-                                errCode = clientException.getMessage();
+                if (activity != null && !activity.isFinishing()) {
+                    mMainHandler.post(() -> {
+                        UIUtils.showToast(R.string.uploading_fail_again);
+                        if (listener != null) {
+                            String errCode = "上传失败";
+                            if (serviceException != null) {
+                                errCode = serviceException.getErrorCode();
+                            } else {
+                                if (clientException != null) {
+                                    errCode = clientException.getMessage();
+                                }
                             }
+                            listener.onFailure(errCode);
                         }
-                        listener.onFailure(errCode);
-                    }
-                });
+                    });
+                }
             }
         });
 
     }
 
 
-    public void uploadingFiles(List<String> filePathData, OnOSSDataResultListener listener, boolean showLoadingView, BaseActivity baseActivity) {
+    public void uploadingFiles(List<String> filePathData, OnOSSDataResultListener listener, boolean showLoadingView, BaseActivity activity) {
         if (mFileData == null) {
             mFileData = checkFiles(filePathData);
             if (mFileData.size() == 0) {
@@ -289,7 +297,7 @@ public class OSSRequestHelp {
                 listener.onStart();
             }
         }
-        if (mFileData.size() == 0){
+        if (mFileData.size() == 0) {
             UIUtils.showToast("上传文件不能为空！");
             return;
         }
@@ -305,52 +313,56 @@ public class OSSRequestHelp {
         request.setProgressCallback((request1, currentSize, totalSize) -> {
 
         });
-        if (showLoadingView){
-            baseActivity.showLoadingView();
+        if (activity != null && !activity.isFinishing() && showLoadingView) {
+            activity.showLoadingView();
         }
         mOssClient.asyncPutObject(request, new OSSCompletedCallback<PutObjectRequest, PutObjectResult>() {
             @Override
             public void onSuccess(PutObjectRequest request, PutObjectResult result) {
-                mMainHandler.post(() -> {
+                if (activity != null && !activity.isFinishing()) {
+                    mMainHandler.post(() -> {
 
-                    String ossFilePath = "https://" + OSSRequestHelp.BUCKET.concat(".").concat(OSSRequestHelp.ENDPOINT_BODY).concat("/")
-                            .concat(request.getObjectKey());
-                    mResultPathData.add(ossFilePath);
-                    filePathData.remove(0);
-                    mFileData.remove(0);
-                    FileUtils.removeFile(file);
-                    if (filePathData.size() == 0) {
-                        if (showLoadingView){
-                            baseActivity.dismissLoadingView();
+                        String ossFilePath = "https://" + OSSRequestHelp.BUCKET.concat(".").concat(OSSRequestHelp.ENDPOINT_BODY).concat("/")
+                                .concat(request.getObjectKey());
+                        mResultPathData.add(ossFilePath);
+                        filePathData.remove(0);
+                        mFileData.remove(0);
+                        FileUtils.removeFile(file);
+                        if (filePathData.size() == 0) {
+                            if (showLoadingView) {
+                                activity.dismissLoadingView();
+                            }
+                            if (listener != null) {
+                                listener.onSucceed(mResultPathData);
+                            }
+                        } else {
+                            uploadingFiles(filePathData, listener, showLoadingView, activity);
                         }
-                        if (listener != null) {
-                            listener.onSucceed(mResultPathData);
-                        }
-                    } else {
-                        uploadingFiles(filePathData, listener);
-                    }
-                });
+                    });
+                }
             }
 
             @Override
             public void onFailure(PutObjectRequest request, ClientException clientException, ServiceException serviceException) {
-                mMainHandler.post(() -> {
-                    if (showLoadingView){
-                        baseActivity.dismissLoadingView();
-                    }
-                    UIUtils.showToast(R.string.uploading_fail_again);
-                    if (listener != null) {
-                        String errCode = "上传失败";
-                        if (serviceException != null) {
-                            errCode = serviceException.getErrorCode();
-                        } else {
-                            if (clientException != null) {
-                                errCode = clientException.getMessage();
-                            }
+                if (activity != null && !activity.isFinishing()) {
+                    mMainHandler.post(() -> {
+                        if (showLoadingView) {
+                            activity.dismissLoadingView();
                         }
-                        listener.onFailure(errCode);
-                    }
-                });
+                        UIUtils.showToast(R.string.uploading_fail_again);
+                        if (listener != null) {
+                            String errCode = "上传失败";
+                            if (serviceException != null) {
+                                errCode = serviceException.getErrorCode();
+                            } else {
+                                if (clientException != null) {
+                                    errCode = clientException.getMessage();
+                                }
+                            }
+                            listener.onFailure(errCode);
+                        }
+                    });
+                }
             }
         });
